Fixing the bug when a file already exists in the folder destination

+++ pc_cleaner.py
@@ -3,10 +3,11 @@ import glob
 import shutil
 
 dict_group={'xlsx':[['*.xlsx','*.xls'],"\\Archivos_XLSX"],'pdf':[['*.pdf'],"\\Archivos_PDF"],'img':[['*.jpg','*.jpeg','*.png','*.gif'],"\\Archivos_IMG"],'videos':[['*.mp4','*.avi','*.mkv','*.wmv'],"\\Archivos_Videos"],'exe':[['*.exe'],"\\Archivos_exe"],'ppt':[['*.pptx'],"\\Archivos_ppt"],'zip':[['*.zip','*.rar','*.tar','*.gz','*.7z'],"\\Archivos_zip"],'txt':[['*.txt','*.csv','*.docx'],"\\Archivos_txt"]}
+list_det_files=['.xlsx','.xls','.pdf','.jpg','.jpeg','.png','.gif','.mp4','.avi','.mkv','.wmv','.exe','.pptx','.zip','.rar','.tar','.gz','.7z','.txt','.csv','.docx']
 clean_path=[]
 single=[]
 flag_path='1'
-method=input('Digita 1 o 2 de pende de que método de limpieza prefiere:\n 1.Sencillo: Escoges una o varios carpetas para organizar y te agrupa los archivos sobrantes en la misma carpeta.\n 2.Avanzado: Escoges ciertas carpetas y digitas ciertos paths, te lo va a origanizar por tipo de archivo ej: si es pdf va a documentos y si es .jpg va a imagenes y así.\n')
+method=input('Digita 1 o 2 depende de qué utilidad desee:\n 1.Sencillo: Escoges una o varios carpetas para organizar y te agrupa los archivos sobrantes en la misma carpeta.\n 2.Por definir.\n')
 while(flag_path=='1'):
     clean_path.append(input('Digita el path de donde se encuentre la carpea que quieras limpiar ej:\n ‪C:\\Users\\Pythonibiris\\OneDrive\\Documents\n').replace('\u202a', ''))
     flag_path=input('Digite 1 o 2 si desea agregar otra carpeta para limpiar? 1.si 2.no\n')
@@ -24,11 +25,23 @@ if method=='1':
                 os.makedirs(path+'\\Otros_archivos')
         for move in dict_group.values():
             for files in move[2]:
-                shutil.move(files, path+move[1])
+                name_file=files.split("\\")[-1]
+                if os.path.exists(path+move[1]+"\\"+name_file):
+                    print("Hey this file: "+ files +" already exists in: "+ path+move[1])
+                else:
+                    shutil.move(files, path+move[1])
         last=glob.glob(os.path.join(path, '*.*'))
         noini = [archivo for archivo in last if 'desktop.ini' not in archivo.lower()]
         for files in last:
-            shutil.move(files, path+'\\Otros_archivos')
+            if any(types in files for types in list_det_files):
+                pass
+            else:        
+                name_file=files.split("\\")[-1]
+                if os.path.exists(path+"Otros_archivos\\"+name_file):
+                    print("Hey this file: "+ files +" already exists in: "+ path+"\\Otros_archivos")
+                else:
+                    shutil.move(files, path+'\\Otros_archivos')
+        input("Press any key to quit the session")
 elif method=='2':
     pass
 else:
